requirements.txt
langchain
langgraph
langchain-openai
python-dotenv
pydantic
requests
numpy